requirements.txt
anyio==3.3.4
APScheduler==3.6.3
asgiref==3.4.1
Authlib==0.15.5
backports.zoneinfo==0.2.1
black==21.9b0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.70.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
h11==0.12.0
httpcore==0.13.7
httptools==0.2.0
httpx==0.20.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mypy-extensions==0.4.3
oauthlib==3.1.1
orjson==3.6.4
pathspec==0.9.0
platformdirs==2.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
pymongo==3.12.0
python-dateutil==2.8.2
python-dotenv==0.19.1
python-jose==3.3.0
python-multipart==0.0.5
python-telegram-bot==13.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
regex==2021.10.8
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rsa==4.7.2
six==1.16.0
slack-sdk==3.11.2
sniffio==1.2.0
starlette==0.16.0
tomli==1.2.1
tornado==6.1
typing-extensions==3.10.0.2
tzdata==2021.3
tzlocal==4.0.1
ujson==4.2.0
urllib3==1.26.7
uvicorn==0.15.0
watchgod==0.7
websockets==10.0
